[#1745] Rename flatbuffer allocator into resizable memory
